Изменил структуру проекта (перенес модели в отдельные файлы)
